split schema-dump into files for each db
